Joining and Alien egg fixes
