+++ UnityProject/Assets/Scripts/Messages/Server/ShootMessage.cs
@@ -32,6 +32,8 @@ public class ShootMessage : ServerMessage {
 	///To be run on client
 	public override void Process()
 	{
+		if (!MatrixManager.IsInitialized) return;
+
 		if (Shooter.Equals(NetId.Invalid)) {
 			//Failfast
 			Logger.LogWarning($"Shoot request invalid, processing stopped: {ToString()}", Category.Firearms);
@@ -40,7 +42,7 @@ public class ShootMessage : ServerMessage {
 
 		//Not even spawned don't show bullets
 		if (PlayerManager.LocalPlayer == null) return;
-		
+
 		LoadMultipleObjects(new uint[] {Shooter, Weapon});
 
 		Gun wep = NetworkObjects[1].GetComponent<Gun>();
